Reorganize code and add unit tests for FMM
